Add a copy of assuming_correctness_of_in (#165)
Rubpicola has a duplicated version of the
Program_logic_goal_for_function tactic which was depending on the tactic
assuming_correctness_of_in.

I am refactoring these tactics in bedrock2. In order to do this easily,
decouple the dependency in rupicola, the tactic it relied on should
probably not have been exposed in this way by bedrock anyways, as it
looks like an implementation detail and not something that should be
public API.

To avoid duplication in the future, one could add the different way to
create program logic goals to bedrock2, and remove
assuming_correctness_of_in and p[rogram_logic_goal_for_function from
rupicola.